# perf(template-compiler): Replace @babel/generator by astring for codegen