--- a/packages/@lwc/compiler/src/__tests__/fixtures/expected-dynamic-component.js
+++ b/packages/@lwc/compiler/src/__tests__/fixtures/expected-dynamic-component.js
@@ -2,10 +2,7 @@ import { registerTemplate, registerComponent, LightningElement, registerDecorato
 import { load } from '@custom/loader';
 
 function tmpl($api, $cmp, $slotset, $ctx) {
-    const {
-        dc: api_dynamic_component,
-        f: api_flatten
-    } = $api;
+    const {dc: api_dynamic_component, f: api_flatten} = $api;
     return api_flatten([api_dynamic_component("x-foo", $cmp.customCtor, {
         context: {
                 lwc: {}