--- a/packages/@lwc/compiler/src/__tests__/fixtures/expected-styled.js
+++ b/packages/@lwc/compiler/src/__tests__/fixtures/expected-styled.js
@@ -5,10 +5,7 @@ return ["\n", (nativeShadow ? ":host {color: blue;}" : [hostSelector, " {color: 
 }
 var _implicitStylesheets = [stylesheet];
 function tmpl($api, $cmp, $slotset, $ctx) {
-const {
-h: api_element,
-c: api_custom_element
-} = $api;
+const {h: api_element, c: api_custom_element} = $api;
 return [api_element("div", {
 key: 0
 }, []), api_custom_element("x-foo", _xFoo, {