--- a/packages/@lwc/compiler/src/__tests__/fixtures/expected-typescript-extension.js
+++ b/packages/@lwc/compiler/src/__tests__/fixtures/expected-typescript-extension.js
@@ -1,9 +1,7 @@
 import { registerTemplate, registerComponent, LightningElement, registerDecorators } from 'lwc';
 
 function tmpl($api, $cmp, $slotset, $ctx) {
-    const {
-        t: api_text
-    } = $api;
+    const {t: api_text} = $api;
     return [api_text("ts test")];
 }