--- a/packages/@lwc/rollup-plugin/src/__tests__/fixtures/expected_default_config_simple_app.js
+++ b/packages/@lwc/rollup-plugin/src/__tests__/fixtures/expected_default_config_simple_app.js
@@ -21,7 +21,6 @@
         key: 0
       }, [api_dynamic($cmp.x)])];
     }
-
     var _tmpl = lwc.registerTemplate(tmpl);
     tmpl.stylesheets = [];
 
@@ -70,7 +69,6 @@
         key: 1
       }, [])])];
     }
-
     var _tmpl$1 = lwc.registerTemplate(tmpl$1);
     tmpl$1.stylesheets = [];
     tmpl$1.stylesheetTokens = {