--- a/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/attribute-allow/expected.js
+++ b/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/attribute-allow/expected.js
@@ -1,5 +1,4 @@
 import { registerTemplate } from "lwc";
-
 function tmpl($api, $cmp, $slotset, $ctx) {
   const { h: api_element } = $api;
   return [
@@ -7,14 +6,13 @@ function tmpl($api, $cmp, $slotset, $ctx) {
       "iframe",
       {
         attrs: {
-          allow: "geolocation https://google-developers.appspot.com"
+          allow: "geolocation https://google-developers.appspot.com",
         },
-        key: 0
+        key: 0,
       },
       []
-    )
+    ),
   ];
 }
-
 export default registerTemplate(tmpl);
 tmpl.stylesheets = [];