--- a/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/attribute-href/expected.js
+++ b/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/attribute-href/expected.js
@@ -1,5 +1,4 @@
 import { registerTemplate } from "lwc";
-
 function tmpl($api, $cmp, $slotset, $ctx) {
   const { t: api_text, h: api_element } = $api;
   return [
@@ -7,25 +6,25 @@ function tmpl($api, $cmp, $slotset, $ctx) {
       "a",
       {
         attrs: {
-          href: "#yasaka-taxi"
+          href: "#yasaka-taxi",
         },
-        key: 0
+        key: 0,
       },
       [api_text("Yasaka Taxi")]
     ),
     api_element(
       "map",
       {
-        key: 1
+        key: 1,
       },
       [
         api_element(
           "area",
           {
             attrs: {
-              href: "#eneos-gas"
+              href: "#eneos-gas",
             },
-            key: 2
+            key: 2,
           },
           []
         ),
@@ -33,16 +32,15 @@ function tmpl($api, $cmp, $slotset, $ctx) {
           "area",
           {
             attrs: {
-              href: "#kawaramachi"
+              href: "#kawaramachi",
             },
-            key: 3
+            key: 3,
           },
           []
-        )
+        ),
       ]
-    )
+    ),
   ];
 }
-
 export default registerTemplate(tmpl);
 tmpl.stylesheets = [];