--- a/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/boolean-attribute/hidden-global-attr/expected.js
+++ b/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/boolean-attribute/hidden-global-attr/expected.js
@@ -1,6 +1,5 @@
 import _xFoo from "x/foo";
 import { registerTemplate } from "lwc";
-
 function tmpl($api, $cmp, $slotset, $ctx) {
   const { t: api_text, h: api_element, c: api_custom_element } = $api;
   return [
@@ -38,7 +37,7 @@ function tmpl($api, $cmp, $slotset, $ctx) {
       "p",
       {
         attrs: {
-          hidden: $cmp.computed ? "" : null,
+          hidden: $cmp.computed ? "" : "null",
         },
         key: 3,
       },
@@ -111,6 +110,5 @@ function tmpl($api, $cmp, $slotset, $ctx) {
     ),
   ];
 }
-
 export default registerTemplate(tmpl);
 tmpl.stylesheets = [];