--- a/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/boolean-attribute/hidden-global-attr/expected.js
+++ b/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/boolean-attribute/hidden-global-attr/expected.js
@@ -37,7 +37,7 @@ function tmpl($api, $cmp, $slotset, $ctx) {
       "p",
       {
         attrs: {
-          hidden: $cmp.computed ? "" : "null",
+          hidden: $cmp.computed ? "" : null,
         },
         key: 3,
       },