--- a/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/boolean-attribute/required/expected.js
+++ b/packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/boolean-attribute/required/expected.js
@@ -1,6 +1,5 @@
 import _xFoo from "x/foo";
 import { registerTemplate } from "lwc";
-
 function tmpl($api, $cmp, $slotset, $ctx) {
   const { h: api_element, t: api_text, c: api_custom_element } = $api;
   return [
@@ -47,7 +46,7 @@ function tmpl($api, $cmp, $slotset, $ctx) {
       "input",
       {
         attrs: {
-          required: $cmp.computed ? "" : null,
+          required: $cmp.computed ? "" : "null",
         },
         props: {
           value: "computed value",
@@ -126,6 +125,5 @@ function tmpl($api, $cmp, $slotset, $ctx) {
     ),
   ];
 }
-
 export default registerTemplate(tmpl);
 tmpl.stylesheets = [];