--- a/packages/@lwc/template-compiler/src/__tests__/fixtures/base/comment/expected.js
+++ b/packages/@lwc/template-compiler/src/__tests__/fixtures/base/comment/expected.js
@@ -1,9 +1,7 @@
 import { registerTemplate } from "lwc";
-
 function tmpl($api, $cmp, $slotset, $ctx) {
   const { t: api_text } = $api;
   return [api_text("Hello world")];
 }
-
 export default registerTemplate(tmpl);
 tmpl.stylesheets = [];