… packages/@lwc/compiler/src/compiler/__tests__/result.spec.ts
… packages/@lwc/compiler/src/rollup-plugins/__tests__/module-resolver.spec.ts
… packages/@lwc/compiler/src/transformers/__tests__/transform-html.spec.ts
… packages/@lwc/template-compiler/package.json
… packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/tabindex-computed/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/class/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/style-expression/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/style-static/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/svg-with-iteration/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/svg/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/template/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/base/text/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-dynamic/flatten-child/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-dynamic/only-child/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/children/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/inline-index/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/inline-item/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/inline-outside-scope-reference/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/inline-refence-item/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/inline-sibling/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/inline/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/template-multiple-children/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/template-outside-scope-reference/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/template-sibling/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-for-each/template/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-if/inline-multiple/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-if/inline-sibling-static/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-if/template-expression/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-if/template-if-else/expected.js
… packages/@lwc/template-compiler/src/__tests__/fixtures/directive-if/template-multiple/expected.js