## Details

This PR replaces `@babel/generator` by `astring` with astring for the template compiler code generation. Astring is a [magnitude faster](https://github.com/davidbonnet/astring/#benchmark) than babel for code generation and extremely well maintained.

Astring doesn't play as well as babel with new lines and whitespaces. Because of this, most of the test baseline had to be regenerated. 

### LWC benchmark

This change results in 15% performance improvement in compiler code evaluation time and 0 to 50% performance speed up in compilation time depending on the use case. The performance gain depends on the presence and size of the HTML template in the LWC module.

**Before:**
```
$ ITERATIONS=20 node ./packages/sfdc-lwc-compiler/scripts/perf-test.js
Compiler loading and evaluation time 3189ms
Warmup...
Measuring average out of 20, component compilation time:
component-small (modes: dev, dev_compat, prod, prod_compat) - 71 ms
component-medium (modes: dev, dev_compat, prod, prod_compat) - 134 ms
component-large (modes: dev, dev_compat, prod, prod_compat) - 178 ms
record-layout-20-fields (modes: dev) - 82 ms
record-layout-60-fields (modes: dev) - 183 ms
record-layout-300-fields (modes: dev) - 876 ms
adg-large (modes: dev) - 227 ms
```

**After:**
```
$ ITERATIONS=20 node ./packages/sfdc-lwc-compiler/scripts/perf-test.js
Compiler loading and evaluation time 2756ms
Warmup...
Measuring average out of 20, component compilation time:
component-small (modes: dev, dev_compat, prod, prod_compat) - 76 ms
component-medium (modes: dev, dev_compat, prod, prod_compat) - 138 ms
component-large (modes: dev, dev_compat, prod, prod_compat) - 186 ms
record-layout-20-fields (modes: dev) - 57 ms
record-layout-60-fields (modes: dev) - 112 ms
record-layout-300-fields (modes: dev) - 543 ms
adg-large (modes: dev) - 226 ms
```

It is still possible to squeeze a 10% performance improvement by replacing the whole AST creation and code generation by direct string generation. It's something that should be evaluated for the next template compiler rewrite. 

## Does this PR introduce breaking changes?

* ✅ `No, it does not introduce breaking changes.`

## The PR fulfills these requirements:
* Have tests for the proposed changes been added? ✅ 
* Have you followed [these instructions](../CONTRIBUTING.md#-commit-message-conventions) to clearly describe the issue being fixed or feature enhanced? ✅ 

## GUS work item
W-7462545